port r2153 (correct r2151)
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/coin@2154 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/coin/src/vandeWiel.c
@@ -281,7 +281,7 @@ void makeW(celW **W, int a, int b, int start, double *rs, double tol) {
     int rank;
     int hulp;
 
-    for (int j = 1; j <= b; j++) {  /* verander naar 0!! */
+    for (long j = 1; j <= b; j++) {  /* verander naar 0!! */
 
         if (j < a) {
             hulp = j;
@@ -289,7 +289,7 @@ void makeW(celW **W, int a, int b, int start, double *rs, double tol) {
             hulp = a;
         }
 
-        for (int i = 1; i <= hulp; i++) {
+        for (long i = 1; i <= hulp; i++) {
             if (i <= j/2 || j == 1) {
                 rank = start+j;
                 fillcell(W, i, j, rank - 1, rs, tol);
